# control/web/docs/swagger.yaml
basePath: /
definitions:
  types.AllPluginStatusParams:
    properties:
      groupId:
        type: integer
      status:
        type: integer
    type: object
  types.PluginStatusParams:
    properties:
      groupId:
        type: integer
      name:
        type: string
      status:
        type: integer
    required:
    - name
    type: object
  types.ResponseStatusParams:
    properties:
      groupId:
        type: integer
      status:
        type: integer
    type: object
  types.SendMsgParams:
    properties:
      gidList:
        items:
          type: integer
        type: array
      message:
        type: string
      selfId:
        type: integer
    type: object
host: 127.0.0.1:3000
info:
  contact: {}
  description: zbp restful api document
  title: zbp api
  version: "1.0"
paths:
  /api/getBotList:
    get:
      description: 获取机器人qq号
      responses: {}
  /api/getFriendList:
    get:
      description: 获取好友列表
      parameters:
      - default: 123456
        description: 机器人qq号
        in: query
        name: selfId
        required: true
        type: integer
      responses: {}
  /api/getGroupList:
    get:
      description: 获取群列表
      parameters:
      - default: 123456
        description: 机器人qq号
        in: query
        name: selfId
        required: true
        type: integer
      responses: {}
  /api/getPermCode:
    get:
      description: 授权码
      responses: {}
  /api/getRequestList:
    get:
      description: 获取所有的请求
      responses: {}
  /api/getUserInfo:
    get:
      description: 获得用户信息
      responses: {}
  /api/handleRequest:
    post:
      description: 处理一个请求
      parameters:
      - default: abc
        description: 事件id
        in: formData
        name: flag
        required: true
        type: string
      - default: abc
        description: 原因
        in: formData
        name: reason
        type: string
      - default: true
        description: 是否同意
        in: formData
        name: approve
        type: boolean
      responses: {}
  /api/login:
    post:
      description: 前端登录
      parameters:
      - default: xiaoguofan
        description: 用户名
        in: formData
        name: username
        required: true
        type: string
      - default: "123456"
        description: 密码
        in: formData
        name: password
        required: true
        type: string
      responses: {}
  /api/logout:
    get:
      description: 登出
      responses: {}
  /api/manage/getAllPlugin:
    get:
      description: 获取所有插件的状态
      parameters:
      - default: 0
        description: 群号
        in: query
        name: groupId
        type: integer
      responses: {}
  /api/manage/getPlugin:
    get:
      description: 获取某个插件的状态
      parameters:
      - default: 0
        description: 群号
        in: query
        name: groupId
        type: integer
      - default: antibuse
        description: 插件名
        in: query
        name: name
        type: string
      responses: {}
  /api/manage/updateAllPluginStatus:
    post:
      description: 更改某群所有插件状态
      parameters:
      - description: 修改插件状态入参
        in: body
        name: object
        schema:
          $ref: '#/definitions/types.AllPluginStatusParams'
      responses: {}
  /api/manage/updatePluginStatus:
    post:
      description: 更改某一个插件状态
      parameters:
      - description: 修改插件状态入参
        in: body
        name: object
        schema:
          $ref: '#/definitions/types.PluginStatusParams'
      responses: {}
  /api/manage/updateResponseStatus:
    post:
      description: 更改某一个群响应
      parameters:
      - description: 修改群响应入参
        in: body
        name: object
        schema:
          $ref: '#/definitions/types.ResponseStatusParams'
      responses: {}
  /api/sendMsg:
    post:
      description: 前端调用发送信息
      parameters:
      - description: 发消息参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/types.SendMsgParams'
      responses: {}
swagger: "2.0"